requirements.txt
openai_whisper==20230314
pyannote.audio==2.1.1
pyannote.core==5.0.0
pyannote.database==5.0.0
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pydub==0.25.1
